+++ app/interactions/api/v1/avatars_controller/shared/upload_avatar.rb
@@ -1,0 +1,12 @@
+class Api::V1::AvatarsController::Shared::UploadAvatar < Api::BaseInteraction
+  object :user
+  object :tempfile
+  string :timestamp
+
+  def execute
+    Aws::S3::Resource.new
+      .bucket(S3Credential::Avatars.instance.cred['bucket'])
+      .object("#{user.mkey}_#{timestamp}")
+      .upload_file(tempfile)
+  end
+end